pyproject.toml
[project]
name = "MangaReader"
version = "0.1.0"
description = "本子浏览器"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
  { name = "kawayiYokami" }
]
license = { file = "LICENSE" }
keywords = ["manga", "reader", "translation", "comic", "image-processing"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics :: Viewers"
]

dependencies = [
    "deep-translator==1.11.4",
    "numpy==2.2.6",
    "opencv-python==4.11.0.86",
    "pillow==11.2.1",
    "pyperclip==1.9.0",
    "pyside6==6.9.0",
    "pyside6-addons==6.9.0",
    "pyside6-essentials==6.9.0",
    "pyside6-fluent-widgets==1.8.1",
    "requests==2.32.3",
    "screeninfo==0.8.1",
    "zhconv==1.4.3",
    "langdetect==1.0.9",
    "flask==3.1.1",
    "gunicorn==23.0.0",
    "opencv-contrib-python==4.11.0.86",
    "onnxruntime-gpu==1.22.0",
    "shapely==2.1.1",
    "pyclipper==1.3.0.post6",
    "scikit-image==0.25.2",
    "imgaug==0.4.0",
    "lmdb==1.6.2",
    "tqdm==4.67.1",
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "jinja2==3.1.6",
    "python-multipart==0.0.20",
    "pywebview==5.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "core",
    "web",
    "web.api",
    "web.websocket",
    "web.utils",
    "utils"
]
